refactor: extract event handler (#41)
