+++ frontend/src/features/skillChaining/skills/components/SkillFlowChart.tsx
@@ -1,4 +1,4 @@
-import { useState, useEffect, useCallback, useMemo, memo } from 'react';
+import { useState, useEffect, useMemo, memo } from 'react';
 import ReactFlow, {
   ConnectionLineType,
   Controls,
@@ -6,7 +6,6 @@ import ReactFlow, {
   useNodesState,
   useEdgesState,
   MiniMap,
-  NodeMouseHandler,
   NodeProps,
   Handle,
   Position,
@@ -17,6 +16,7 @@ import { LoadingIndicator } from '@components/common/LoadingIndicator';
 import { ErrorMessage } from '@components/common/ErrorMessage';
 import { getCategoryColor } from '@features/skillChaining/categories/hooks/categoryColors';
 import { useSkillGraph } from '../hooks/useSkillGraph';
+import { useGraphInteractions } from '../hooks/useGraphInteractions';
 import { SkillNodeData } from '../utils/skillGraphUtils';
 import './SkillFlowChart.css';
 
@@ -70,6 +70,16 @@ export function SkillFlowChart({
   const [edges, setEdges, onEdgesChange] = useEdgesState([]);
   const [highlightedNodes, setHighlightedNodes] = useState<string[]>([]);
   
+  // グラフインタラクション用のカスタムフック
+  const { onNodeMouseEnter, onNodeMouseLeave, onNodeClick } = useGraphInteractions({
+    edges,
+    sourceSkillName,
+    setNodes,
+    setEdges,
+    setHighlightedNodes,
+    onSkillSelect
+  });
+  
   // グラフデータが更新されたらReactFlowの状態を更新
   useEffect(() => {
     if (loading || error) {
@@ -86,90 +96,6 @@ export function SkillFlowChart({
     }, 0);
     setHighlightedNodes([sourceSkillName]);
   }, [graphNodes, graphEdges, sourceSkillName, loading, error, setNodes, setEdges]);
-  
-  // ノードホバー時のハイライト処理
-  const onNodeMouseEnter: NodeMouseHandler = useCallback((_, node) => {
-    try {
-      // ホバーしたノードと直接関連するノードをハイライト
-      const connectedEdges = edges.filter(e => 
-        e.source === node.id || e.target === node.id
-      );
-      
-      const relatedNodeIds = [
-        node.id,
-        ...connectedEdges.map(e => e.source === node.id ? e.target : e.source)
-      ];
-      
-      setHighlightedNodes(relatedNodeIds);
-      
-      // 関連ノードの強調表示（カスタムノード用に修正）
-      setNodes(nds => 
-        nds.map(n => ({
-          ...n,
-          // ノードクラス名でハイライト状態を制御
-          className: relatedNodeIds.includes(n.id) ? 'highlighted-node' : 'faded-node'
-        }))
-      );
-      
-      // 関連エッジの強調表示
-      setEdges(eds => 
-        eds.map(e => ({
-          ...e,
-          animated: connectedEdges.some(ce => ce.id === e.id),
-          style: {
-            ...e.style,
-            opacity: connectedEdges.some(ce => ce.id === e.id) ? 1 : 0.1,
-            zIndex: connectedEdges.some(ce => ce.id === e.id) ? 1 : 0
-          }
-        }))
-      );
-    } catch (error) {
-      console.error('ノードホバー処理でエラーが発生しました:', error);
-    }
-  }, [edges, setNodes, setEdges]);
-  
-  // ノードホバー解除時の処理
-  const onNodeMouseLeave = useCallback(() => {
-    // 全てのノードとエッジを元の表示に戻す
-    setNodes(nds => 
-      nds.map(n => ({
-        ...n,
-        className: '', // クラス名をリセット
-      }))
-    );
-    
-    setEdges(eds => 
-      eds.map(e => ({
-        ...e,
-        style: {
-          ...e.style,
-          opacity: 1,
-          zIndex: 0
-        }
-      }))
-    );
-    
-    setHighlightedNodes([sourceSkillName]);
-  }, [sourceSkillName, setNodes, setEdges]);
-  
-  // ノードクリック時の処理
-  const onNodeClick: NodeMouseHandler = useCallback((_, node) => {
-    try {
-      // ノードデータからスキル名を取得（IDからではなく）
-      const clickedSkillName = node.data?.label;
-      if (typeof clickedSkillName === 'string') {
-        
-        // ノードのIDに「source_」が含まれているかで判断
-        const shouldAddToChain = !node.id.toString().startsWith('source_');
-        
-        if (onSkillSelect) {
-          onSkillSelect(clickedSkillName, shouldAddToChain);
-        }
-      }
-    } catch (error) {
-      console.error('ノードクリック処理でエラーが発生しました:', error);
-    }
-  }, [onSkillSelect]);
   
   if (loading) return <div className="loading-container"><LoadingIndicator /></div>;
   if (error) return <div className="error-container"><ErrorMessage message={error?.message || 'エラーが発生しました'} /></div>;
